+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## _(in-progress)_
 * Change: Switch to using `wp_rand()` for more reliable randomization
 * Change: Allow `$extensions` argument to also accept an array of extensions
+* Change: Strip surrounding whitespace and leading periods from provided extensions
 * Change: Explicitly return false if no file could be found
 * Change: Switch to use a switch statement, which is more concise
 * Change: Use cleaner `sprintf()` approach to outputting markup